+++ star.js
@@ -45,9 +45,11 @@ cmd
   .option('    --code [code]'  , 'specify the stock code of insider tradings you want to query. data source: traceinvest.com')
   .option('    --market <mkt>' , 'specify the market of insider trading query, case insensitive:SZM-深圳主板, SZGEM-深圳创业板, SZSME-深圳中小板,\n' + ' '.repeat(20) +
                                  ' SHM-上海主板. multiple market should be separated by "," or "，". ')
-  .option('--latest-sz'        , 'query latest insider tradings of ShenZhen market, should be used with "-i" or "--insider". data source: szse')
-  .option('--latest-sh'        , 'query latest insider tradings of ShangHai market, should be used with "-i" or "--insider". data source: sse')
-  .option('--show-detail'      , 'show detail latest insider trading records, should be used with "-i" or "--insider".')
+  .option('    --top-buy'      , 'query top buy of insider tradings, should be used with "-i" or "--insider". time span:1m~12m.')
+  .option('    --top-sell'     , 'query top sell of insider tradings, should be used with "-i" or "--insider". time span:1m~12m.')
+  .option('    --latest-sz'    , 'query latest insider tradings of ShenZhen market, should be used with "-i" or "--insider". data source: szse')
+  .option('    --latest-sh'    , 'query latest insider tradings of ShangHai market, should be used with "-i" or "--insider". data source: sse')
+  .option('    --show-detail'  , 'show detail latest insider trading records, should be used with "-i" or "--insider".')
   .option('-w, --watch [c1...]', 'watch specified stocks or watch all the stocks in watch list.')
   .option('-r, --reverse'      , 'sort stocks in ascending order according to designated field.')
   .option('-l, --limit <n>'    , 'set total display limit of current page.', parseInt)
@@ -93,8 +95,10 @@ let actions = {
         let async   = require('async');
         let Insider = require('./lib/insider.js').Insider;
 
-        if(cmd.latestSz){ Insider.querySZLatest(); return false; }
-        if(cmd.latestSh){ Insider.querySHLatest(); return false; }
+        if(cmd.latestSz){ Insider.querySZLatest();    return false; }
+        if(cmd.latestSh){ Insider.querySHLatest();    return false; }
+        if(cmd.topBuy)  { Insider.queryTopList('BV'); return false; }
+        if(cmd.topSell) { Insider.queryTopList('SV'); return false; }
         if(cmd.insider === true){ Insider.queryMiscInsider(); return false; }
 
         let query   = cmd.insider.replace(/，/g, ',');